mc/requirements.txt
python-dotenv==1.1.0
openai==1.82.0
google-generativeai==0.8.5
together==1.5.13